requirements.txt
pytesseract
Pillow
sentence-transformers
chromadb
packaging
transformers
torch
numpy
scikit-learn
pandas